fix: deploy failure caused by incorrect date type
Previously, output format of datetime widget defined in config.yml
caused the date field of the frontmatter to be of type string,
instead of the desired Date object which caused deploy failure due to
the invalid GraphQL query calling 'formatString' on a string field.
This commit fixes that.

Additionally, as of this commit there are certain things to take note of
when it comes to NetlifyCMS. First, the relative path for the images in
the post body has to be manually set due to an error within
gatsby-remark-relative-images which will be fixed by the following PR:
https://github.com/danielmahon/gatsby-remark-relative-images/pull/30.
Second, the preview for the images in the body doesn't work because the
image path gets directly inserted as the src of an img tag in the
client-side MDX renderer instead of being transformed to the appropriate
path (which is most likely handled by gatsby-plugin-mdx)